+++ RELEASE_NOTES.md
@@ -10,6 +10,8 @@ Also, as of this release, Java 11 is required.
 - `Removed` method Node.remove().
 - `Removed` default method Node.isLeaf().
 - `Removed` method AbstractNode.isLeaf().
+- `Removed` SDAParseException().
+- `Changed` be.baur.sda.io.ParseException() is no longer abstract.
 
 ## Compatibility